vulkan h265 /w 10-bit + fix encoder hang (#565)

+++ server/encoder/video_encoder_vulkan_h264.cpp
@@ -16,6 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 #include "video_encoder_vulkan_h264.h"
+
 #include "encoder/encoder_settings.h"
 #include "utils/wivrn_vk_bundle.h"
 
@@ -337,7 +338,7 @@ void * wivrn::video_encoder_vulkan_h264::encode_info_next(uint32_t frame_num, si
 	std_picture_info = {
 	        .flags =
 	                {
-	                        .IdrPicFlag = uint32_t(ref_slot ? 0 : 1),
+	                        .IdrPicFlag = ref_slot ? 0u : 1u,
 	                        .is_reference = 1,
 	                        .no_output_of_prior_pics_flag = 0,
 	                        .long_term_reference_flag = 0,
